[CI] Make MSCV artifacts output match the regular artifact CI
Merges the two separate archives in to one and puts the dll and pdb files in a x32 and x64 folder.
Currently the two archives package the output in a folder called src and the files in subfolders of the respective names. E.g. `src/d3d9/d3d9.dll` etc.

This makes the MSCV artifacts archive match the regular artifacts CI output.